# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AndMI API/Open API System",
        "description": "Andijon Mashinasozlik Instituti Ochiq API tizimi",
        "contact": {
            "email": "andmiatmdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/formats": {
            "get": {
                "tags": [
                    "Форматы"
                ],
                "summary": "Получение списка всех форматов",
                "operationId": "5129d9264b4b98c592e9adf2a58f699f",
                "responses": {
                    "200": {
                        "description": "Список форматов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "PDF"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Форматы"
                ],
                "summary": "Создание формата",
                "operationId": "de9f4bbb096c06fb82231b164bec0955",
                "requestBody": {
                    "description": "Данные для создания формата",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PDF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Формат успешно создан"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/formats/{id}": {
            "get": {
                "tags": [
                    "Форматы"
                ],
                "summary": "Получение формата",
                "operationId": "94300558ebbd200a4639dd520b0f93bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID формата",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Формат найден"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Формат не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Форматы"
                ],
                "summary": "Обновление формата",
                "operationId": "be601505f007d8630d5077aa5f91ba28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID формата",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления формата",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PDF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Формат успешно обновлен"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Формат не найден"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Форматы"
                ],
                "summary": "Удаление формата",
                "operationId": "632a4af8ce256a92943cdacac678637e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID формата",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Формат успешно удален"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Формат не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token orqali autentifikatsiyadan o'tish",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}